requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
discord.py==2.3.1
elastic-transport==8.4.0
elasticsearch==7.17.9
fonttools==4.42.1
frozenlist==1.3.3
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
importlib-resources==6.0.1
instaloader==4.9
kiwisolver==1.4.5
matplotlib==3.7.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.22.4
openai==0.27.8
packaging==23.1
Pillow==9.1.1
pkg_resources==0.0.0
proglog==0.1.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
yarl==1.8.2
zipp==3.16.2
